Move ReanimatedError to common
