+++ packages/react-native-reanimated/src/animation/transformationMatrix/matrixUtils.tsx
@@ -1,6 +1,6 @@
 'use strict';
 
-import { ReanimatedError } from '../../errors';
+import { ReanimatedError } from '../../common';
 
 type FixedLengthArray<
   T,
