+++ packages/react-native-reanimated/src/animation/util.ts
@@ -17,6 +17,7 @@ import {
   toGammaSpace,
   toLinearSpace,
 } from '../Colors';
+import { ReanimatedError } from '../common';
 import type {
   AnimatableValue,
   AnimatableValueObject,
@@ -28,7 +29,6 @@ import type {
 } from '../commonTypes';
 import { ReduceMotion } from '../commonTypes';
 import type { EasingFunctionFactory } from '../Easing';
-import { ReanimatedError } from '../errors';
 import { shouldBeUseWeb } from '../PlatformChecker';
 import { ReducedMotionManager } from '../ReducedMotion';
 import type { HigherOrderAnimation, StyleLayoutAnimation } from './commonTypes';
