+++ packages/react-native-reanimated/src/common/index.ts
@@ -1,5 +1,6 @@
 'use strict';
 export * from './constants';
+export * from './errors';
 export * from './processors';
 export * from './types';
 export * from './utils';
