+++ packages/react-native-reanimated/src/core.ts
@@ -6,6 +6,7 @@ import {
 import type { WorkletFunction } from 'react-native-worklets';
 import { makeShareableCloneRecursive } from 'react-native-worklets';
 
+import { ReanimatedError } from './common';
 import type {
   AnimatedKeyboardOptions,
   LayoutAnimationBatchItem,
@@ -15,7 +16,6 @@ import type {
   Value3D,
   ValueRotation,
 } from './commonTypes';
-import { ReanimatedError } from './errors';
 import { shouldBeUseWeb } from './PlatformChecker';
 import { ReanimatedModule } from './ReanimatedModule';
 import { SensorContainer } from './SensorContainer';
