+++ packages/react-native-reanimated/src/css/easings/cubicBezier.ts
@@ -1,5 +1,5 @@
 'use strict';
-import { ReanimatedError } from '../errors';
+import { ReanimatedError } from '../../common';
 import type {
   NormalizedCubicBezierEasing,
   ParametrizedTimingFunction,
