+++ packages/react-native-reanimated/src/css/easings/steps.ts
@@ -1,5 +1,5 @@
 'use strict';
-import { ReanimatedError } from '../errors';
+import { ReanimatedError } from '../../common';
 import type { Point } from '../types';
 import type {
   NormalizedStepsEasing,
