+++ packages/react-native-reanimated/src/css/platform/native/normalization/animation/__tests__/settings.test.ts
@@ -1,6 +1,6 @@
 'use strict';
+import { ReanimatedError } from '../../../../../../common';
 import { cubicBezier } from '../../../../../easings';
-import { ReanimatedError } from '../../../../../errors';
 import type {
   CSSAnimationDirection,
   CSSAnimationFillMode,
