+++ packages/react-native-reanimated/src/css/platform/native/normalization/animation/settings.ts
@@ -1,5 +1,5 @@
 'use strict';
-import { ReanimatedError } from '../../../../errors';
+import { ReanimatedError } from '../../../../../common';
 import type {
   CSSAnimationDirection,
   CSSAnimationFillMode,
