+++ packages/react-native-reanimated/src/css/platform/native/normalization/transition/shorthand.ts
@@ -1,7 +1,7 @@
 'use strict';
+import { ReanimatedError } from '../../../../../common';
 import type { ControlPoint, CSSTimingFunction } from '../../../../easings';
 import { cubicBezier, linear, steps } from '../../../../easings';
-import { ReanimatedError } from '../../../../errors';
 import type {
   ConvertValuesToArraysWithUndefined,
   CSSTransitionProperties,
