+++ packages/react-native-reanimated/src/css/platform/native/style/processors/__tests__/colors.test.ts
@@ -1,5 +1,5 @@
 'use strict';
-import { ReanimatedError } from '../../../../../errors';
+import { ReanimatedError } from '../../../../../../common';
 import { ERROR_MESSAGES, processColor } from '../colors';
 
 describe(processColor, () => {
