+++ packages/react-native-reanimated/src/css/platform/native/style/processors/__tests__/transform.test.ts
@@ -1,5 +1,5 @@
 'use strict';
-import { ReanimatedError } from '../../../../../errors';
+import { ReanimatedError } from '../../../../../../common';
 import type { TransformsArray } from '../../../../../types';
 import { ERROR_MESSAGES, processTransform } from '../transform';
