+++ packages/react-native-reanimated/src/css/platform/native/style/processors/__tests__/transformOrigin.test.ts
@@ -6,8 +6,8 @@ import type {
 import {
   ERROR_MESSAGES,
   processTransformOrigin,
+  ReanimatedError,
 } from '../../../../../../common';
-import { ReanimatedError } from '../../../../../errors';
 
 describe(processTransformOrigin, () => {
   describe('valid cases', () => {
