+++ packages/react-native-reanimated/src/css/platform/web/utils.ts
@@ -2,9 +2,9 @@
 import type { ColorValue, DimensionValue } from 'react-native';
 
 import { processColor } from '../../../Colors';
+import { ReanimatedError } from '../../../common';
 import type { ParametrizedTimingFunction } from '../../easings';
 import { CubicBezierEasing, LinearEasing, StepsEasing } from '../../easings';
-import { ReanimatedError } from '../../errors';
 import type { AddArrayPropertyType, ConvertValuesToArrays } from '../../types';
 import { kebabizeCamelCase } from '../../utils';
