+++ packages/react-native-reanimated/src/fabricUtils.web.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-import { ReanimatedError } from './errors';
+import { ReanimatedError } from './common';
 
 export function getShadowNodeWrapperFromRef() {
   throw new ReanimatedError(
