+++ packages/react-native-reanimated/src/hook/useAnimatedStyle.ts
@@ -5,6 +5,7 @@ import type { WorkletFunction } from 'react-native-worklets';
 import { isWorkletFunction } from 'react-native-worklets';
 
 import { initialUpdaterRun } from '../animation';
+import { ReanimatedError } from '../common';
 import { processBoxShadow } from '../common/processors';
 import type {
   AnimatedPropsAdapterFunction,
@@ -18,7 +19,6 @@ import type {
 } from '../commonTypes';
 import { makeShareable, startMapper, stopMapper } from '../core';
 import type { AnimatedProps } from '../createAnimatedComponent/commonTypes';
-import { ReanimatedError } from '../errors';
 import { isJest, shouldBeUseWeb } from '../PlatformChecker';
 import { updateProps, updatePropsJestWrapper } from '../updateProps';
 import type { ViewDescriptorsSet } from '../ViewDescriptorsSet';
