+++ packages/react-native-reanimated/src/hook/utils.ts
@@ -1,7 +1,7 @@
 'use strict';
 import type { WorkletFunction } from 'react-native-worklets';
 
-import { ReanimatedError } from '../errors';
+import { ReanimatedError } from '../common';
 import type { DependencyList } from './commonTypes';
 
 // Builds one big hash from multiple worklets' hashes.
